MatHax Integration
MatHax is a fork of Meteor, and as such most of its codebase is very similar to that of Meteor's, making implementing support for it very easy now that I have Meteor support implemented.

However, MatHax removed support for add-ons, which made hooking into it somewhat tricky as there wasn't a "good" way of figuring out when it initializes. I worked around this by simply using a mixin on MatHax's main class, but good god is that a hackjob.