Do not allow Minecraft to set java.awt.headless to true
For whatever reason, Mojang thought it was a spectacular idea to set java.awt.headless to false in the clinit of net.minecraft.client.main.Main.class.
Perhaps it increases performance? But that doesn't particularly matter in our case, as we need to use AWT for the Blackbox module.
I also removed the headless set that Blackbox automatically does when you try to open it, as this fix makes it redundant.
Thanks to @LunaWasFlaggedAgain for pointing this out.
Fixes #18